change to math local eval
